+++ src/vscode/notebook/_types.ts
@@ -1,0 +1,7 @@
+export interface DatabricksLanguageMapping { 
+	databricksLanguage: string, 
+	vscodeLanguage: string, 
+	magic: string, 
+	commentCharacters: string, 
+	fileExtension: string 
+}